separated FieldConstants
